+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Created and configured the helper functions & component library project.
 - Implemented a Badge component.
 - Implemented a Button component.
+- Implemented a Card component.
 - Implemented a Checkbox component.
 - Implemented a Progress component.
 - Implemented a Select component.